refactor: update folder system
