+++ src/pages/ErrorPage.tsx
@@ -1,6 +1,6 @@
 import { Link, useRouteError } from 'react-router-dom';
-import Alert from '../components/Alert';
-import Button from '../components/Button';
+import Alert from '../components/common/Alert';
+import Button from '../components/common/Button';
 
 const ErrorPage = () => {
   const error = useRouteError() as Error;
